# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "GlazeAudio API",
        "description": "REST API for GlazeAudio \u2013 a platform for reviewing music. Resources are hierarchical: albums \u2192 songs \u2192 reviews. Reviews are written by users.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:5080/"
        }
    ],
    "paths": {
...
        "/api/albums/{albumId}/songs/{songId}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "List reviews of a song (paged, filterable)",
                "description": "Returns one page of the song's reviews, newest first. Filter with author (username), userId, minRating and maxRating (overall rating = average of the 4 aspects). 404 if the album or song does not exist, or the song is not on that album.",
                "operationId": "GetReviews",
                "parameters": [
                    {
                        "name": "albumId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "songId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number, starting at 1. Default 1.",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Items per page, 1\u201350. Default 10.",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Only reviews by this author's username (exact match, case-insensitive).",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "Only reviews written by this user id.",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "minRating",
                        "in": "query",
                        "description": "Only reviews whose overall rating (average of the 4 aspects) is at least this value, 0\u20135.",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "maxRating",
                        "in": "query",
                        "description": "Only reviews whose overall rating is at most this value, 0\u20135.",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedResultOfReviewDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpValidationProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Review a song",
                "description": "Each aspect (lyrics, melody, mood, expressiveness) is rated 0\u20135. The author is given as userId (until authentication is added). 422 if a rating is out of range or the user does not exist; 409 if this user has already reviewed the song.",
                "operationId": "CreateReview",
                "parameters": [
                    {
                        "name": "albumId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "songId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateReviewRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpValidationProblemDetails"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/albums/{albumId}/songs/{songId}/reviews/{reviewId}": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get one review",
                "operationId": "GetReview",
                "parameters": [
                    {
                        "name": "albumId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "songId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "reviewId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Edit a review",
                "description": "Replaces the ratings and comment. The author cannot be changed.",
                "operationId": "UpdateReview",
                "parameters": [
                    {
                        "name": "albumId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "songId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "reviewId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateReviewRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpValidationProblemDetails"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete a review",
                "description": "Used by administrators to remove harmful reviews. Returns 204 with no body.",
                "operationId": "DeleteReview",
                "parameters": [
                    {
                        "name": "albumId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "songId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "reviewId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users/{userId}/reviews": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List a user's reviews (paged)",
                "description": "Every review the user has written, newest first, with the song and album each one belongs to. Filter with minRating.",
                "operationId": "GetUserReviews",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number, starting at 1. Default 1.",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Items per page, 1\u201350. Default 10.",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "minRating",
                        "in": "query",
                        "description": "Only reviews whose overall rating is at least this value, 0\u20135.",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedResultOfUserReviewDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpValidationProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateReviewRequest": {
                "required": [
                    "userId",
                    "lyricsRating",
                    "melodyRating",
                    "moodRating",
                    "expressivenessRating",
                    "comment"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "maximum": 2147483647,
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "lyricsRating": {
                        "maximum": 5,
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "melodyRating": {
                        "maximum": 5,
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "moodRating": {
                        "maximum": 5,
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "expressivenessRating": {
                        "maximum": 5,
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "comment": {
                        "maxLength": 2000,
                        "minLength": 1,
                        "type": "string",
                        "nullable": true
                    }
                }
            },
...
            "HttpValidationProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "detail": {
                        "type": "string",
                        "nullable": true
                    },
                    "instance": {
                        "type": "string",
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
...
            "PagedResultOfReviewDto": {
                "required": [
                    "items",
                    "page",
                    "pageSize",
                    "totalCount",
                    "totalPages"
                ],
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReviewDto"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "_links": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            },
...
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "detail": {
                        "type": "string",
                        "nullable": true
                    },
                    "instance": {
                        "type": "string",
                        "nullable": true
                    }
                }
            },
...
            "UpdateReviewRequest": {
                "required": [
                    "lyricsRating",
                    "melodyRating",
                    "moodRating",
                    "expressivenessRating",
                    "comment"
                ],
                "type": "object",
                "properties": {
                    "lyricsRating": {
                        "maximum": 5,
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "melodyRating": {
                        "maximum": 5,
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "moodRating": {
                        "maximum": 5,
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "expressivenessRating": {
                        "maximum": 5,
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "comment": {
                        "maxLength": 2000,
                        "minLength": 1,
                        "type": "string",
                        "nullable": true
                    }
                }
            },
...
        }
    },
...
}